requirements.txt
streamlit
torch
torchvision
torchaudio
diffusers
transformers
accelerate
xformers
Pillow
